# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Code Review Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface GithubAuthRequest {
  code: string;
  redirectUri?: string;
}

export interface HealthStatus {
  status: string;
}

export interface Repository {
  id: number;
  name: string;
  fullName: string;
  language: string;
  stars: number;
  openPrs: number;
  /** @nullable */
  lastReviewedAt: string | null;
  /** @nullable */
  reviewScore: number | null;
  /** @nullable */
  description?: string | null;
  isPrivate?: boolean;
}

export type PullRequestStatus = typeof PullRequestStatus[keyof typeof PullRequestStatus];


export const PullRequestStatus = {
  open: 'open',
  closed: 'closed',
  merged: 'merged',
} as const;

/**
 * @nullable
 */
export type PullRequestReviewStatus = typeof PullRequestReviewStatus[keyof typeof PullRequestReviewStatus] | null;


export const PullRequestReviewStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  completed: 'completed',
} as const;

export interface PullRequest {
  id: number;
  number: number;
  title: string;
  author: string;
  /** @nullable */
  authorAvatar?: string | null;
  repositoryId: number;
  repositoryName: string;
  filesChanged: number;
  commits: number;
  status: PullRequestStatus;
  /** @nullable */
  reviewStatus?: PullRequestReviewStatus;
  createdAt: string;
}

export type AgentStatusStatus = typeof AgentStatusStatus[keyof typeof AgentStatusStatus];


export const AgentStatusStatus = {
  waiting: 'waiting',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface AgentStatus {
  name: string;
  status: AgentStatusStatus;
  /** @nullable */
  progress?: number | null;
}

export interface ReviewJob {
  jobId: string;
  pullRequestId: number;
  status: string;
  progress?: number;
  agents?: AgentStatus[];
  reviewId?: number;
  error?: string;
}

export interface Review {
  id: number;
  pullRequestId: number;
  pullRequestNumber: number;
  repositoryName: string;
  author: string;
  /** @nullable */
  authorAvatar?: string | null;
  qualityScore: number;
  securityScore: number;
  complexityScore: number;
  documentationScore: number;
  status: string;
  reviewedAt: string;
  /** @nullable */
  aiSummary?: string | null;
  topIssues?: string[];
  /** @nullable */
  modelUsed?: string | null;
  /** @nullable */
  inputTokens?: number | null;
  /** @nullable */
  outputTokens?: number | null;
  /** @nullable */
  cost?: number | null;
  /** @nullable */
  agentOutputs?: string | null;
  /** @nullable */
  promptTemplates?: string | null;
  /** @nullable */
  latencyMs?: number | null;
}

export type ReviewFileStatus = typeof ReviewFileStatus[keyof typeof ReviewFileStatus];


export const ReviewFileStatus = {
  added: 'added',
  modified: 'modified',
  deleted: 'deleted',
} as const;

export interface ReviewFile {
  id: number;
  reviewId: number;
  filePath: string;
  additions: number;
  deletions: number;
  status: ReviewFileStatus;
  /** @nullable */
  content?: string | null;
}

export type ReviewCommentType = typeof ReviewCommentType[keyof typeof ReviewCommentType];


export const ReviewCommentType = {
  bug: 'bug',
  security: 'security',
  complexity: 'complexity',
  documentation: 'documentation',
  style: 'style',
  performance: 'performance',
} as const;

export type ReviewCommentSeverity = typeof ReviewCommentSeverity[keyof typeof ReviewCommentSeverity];


export const ReviewCommentSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface ReviewComment {
  id: number;
  reviewId: number;
  filePath: string;
  line: number;
  type: ReviewCommentType;
  severity: ReviewCommentSeverity;
  message: string;
  /** @nullable */
  suggestion?: string | null;
  /** @nullable */
  codeSnippet?: string | null;
}

export type SecurityIssueSeverity = typeof SecurityIssueSeverity[keyof typeof SecurityIssueSeverity];


export const SecurityIssueSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type SecurityIssueStatus = typeof SecurityIssueStatus[keyof typeof SecurityIssueStatus];


export const SecurityIssueStatus = {
  open: 'open',
  fixed: 'fixed',
  ignored: 'ignored',
} as const;

export interface SecurityIssue {
  id: number;
  type: string;
  filePath: string;
  line: number;
  severity: SecurityIssueSeverity;
  description: string;
  /** @nullable */
  fix?: string | null;
  status: SecurityIssueStatus;
  /** @nullable */
  repositoryId?: number | null;
}

export interface SecuritySummary {
  critical: number;
  high: number;
  medium: number;
  low: number;
  total: number;
}

export interface DashboardStats {
  totalRepositories: number;
  totalPullRequests: number;
  filesReviewed: number;
  overallScore: number;
  activeReviews: number;
  reviewsThisWeek?: number;
  issuesFound?: number;
}

export interface TrendPoint {
  date: string;
  qualityScore: number;
  securityScore: number;
  complexityScore?: number;
}

export interface LanguageStat {
  language: string;
  percentage: number;
  count: number;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  githubUsername: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  timezone?: string | null;
  joinedAt: string;
  totalReviews?: number;
  totalRepositories?: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  review_completed: 'review_completed',
  pr_analyzed: 'pr_analyzed',
  security_found: 'security_found',
  repository_added: 'repository_added',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  repositoryName?: string | null;
  createdAt: string;
}

export interface ReviewHistoryItem {
  id: number;
  pullRequestNumber: number;
  repositoryName: string;
  score: number;
  status: string;
  reviewedAt: string;
  /** @nullable */
  author?: string | null;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
  rememberMe?: boolean;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface SuccessResponse {
  success: boolean;
}

export type ListRepositoriesParams = {
language?: string;
search?: string;
};

export type ListPullRequestsParams = {
status?: ListPullRequestsStatus;
repositoryId?: number;
};

export type ListPullRequestsStatus = typeof ListPullRequestsStatus[keyof typeof ListPullRequestsStatus];


export const ListPullRequestsStatus = {
  open: 'open',
  closed: 'closed',
  merged: 'merged',
} as const;

export type ListReviewsParams = {
repositoryId?: number;
limit?: number;
};

export type ListSecurityIssuesParams = {
severity?: ListSecurityIssuesSeverity;
repositoryId?: number;
};

export type ListSecurityIssuesSeverity = typeof ListSecurityIssuesSeverity[keyof typeof ListSecurityIssuesSeverity];


export const ListSecurityIssuesSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type GetQualityTrendParams = {
days?: number;
};

export type GetReviewHistoryParams = {
search?: string;
limit?: number;
};