# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Code Review Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/github:
    post:
      operationId: githubAuth
      tags: [auth]
      summary: Authenticate via GitHub OAuth code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubAuthRequest"
      responses:
        "200":
          description: GitHub authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /repositories:
    get:
      operationId: listRepositories
      tags: [repositories]
      summary: List all repositories
      parameters:
        - name: language
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repository"
...
  /analytics/quality-trend:
    get:
      operationId: getQualityTrend
      tags: [analytics]
      summary: Get code quality trend over time
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Quality trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"
...
components:
  schemas:
    GithubAuthRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
        redirectUri:
          type: string
...
    Repository:
      type: object
      required: [id, name, fullName, language, stars, openPrs, lastReviewedAt, reviewScore]
      properties:
        id:
          type: integer
        name:
          type: string
        fullName:
          type: string
        language:
          type: string
        stars:
          type: integer
        openPrs:
          type: integer
        lastReviewedAt:
          type: ["string", "null"]
        reviewScore:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        isPrivate:
          type: boolean
...
    TrendPoint:
      type: object
      required: [date, qualityScore, securityScore]
      properties:
        date:
          type: string
        qualityScore:
          type: number
        securityScore:
          type: number
        complexityScore:
          type: number
...
    UserProfile:
      type: object
      required: [id, name, email, githubUsername, avatarUrl, joinedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        githubUsername:
          type: string
        avatarUrl:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        joinedAt:
          type: string
        totalReviews:
          type: integer
        totalRepositories:
          type: integer
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
...